More work on getting rid of warnings the Right Way

+++ src/TipOfTheDay.cpp
@@ -201,7 +201,7 @@ TipOfTheDay::TipOfTheDay(wxWindow *parent)
   config->Read(wxT("tipNum"), &m_num);
   if(m_num < 0)
     m_num = m_tips.GetCount()-1;
-  if((unsigned)m_num >=m_tips.GetCount())
+  if(m_num >=m_tips.GetCount())
     m_num = 0;
 
   SetName("TipOfTheDay");
@@ -329,7 +329,7 @@ wxImage TipOfTheDay::GetImage(unsigned char *data, size_t len)
 void TipOfTheDay::OnNextButton(wxCommandEvent &WXUNUSED(dummy))
 {
   m_num++;
-  if((unsigned)m_num >=m_tips.GetCount())
+  if(m_num >=m_tips.GetCount())
     m_num = 0;
   m_tip->SetValue(m_tips[m_num]);
 }
